deps/v8/test/inspector/regress/regress-crbug-1220203-expected.txt

Regression test for crbug.com/1220203.

Running test: testBreakOnUncaughtException
Uncaught exception at
function throwError() {
  #throw new Error();
}
